examples/requirements-txt/requirements.txt
flask>=3.0
gunicorn>=21.2
redis>=4.5
python-dotenv>=1.0
